% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D200TechSimilarityTask.R
\name{D200TechSimilarityTask}
\alias{D200TechSimilarityTask}
\title{Create a list of corpus similarity based recommendations.}
\description{
Imagine this object is a lab. You can use two methods in this lab
easy to make a recommendation list.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{raw_corpus}}{Preprocessed corpus after initialize.}

\item{\code{corpus_matrix}}{The output of the private method
\code{get_vectorize_matrix}.}

\item{\code{insight_matrix_set}}{The output of the public method
\code{insight_matrix_set} ( the raw result of similarity between
\code{main_item} and \code{ref_item} ).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{D200TechSimilarityTask$new()}}
\item \href{#method-get_insight_matrix}{\code{D200TechSimilarityTask$get_insight_matrix()}}
\item \href{#method-get_recommend_result}{\code{D200TechSimilarityTask$get_recommend_result()}}
\item \href{#method-get_connection}{\code{D200TechSimilarityTask$get_connection()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new D200TechSimilarityTask object. You
can image this step is to create or enter into the lab you can
easy to make a recommendation list.
This initializing step also do the following work :
\enumerate{
\item Check column name.
\item Check data type (tibble).
\item If the data type is wrong, it will auto convert
to the right data type.
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{D200TechSimilarityTask$new(raw_corpus)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{raw_corpus}}{The raw corpus you want to analysis
from external place(.csv or .xlsx file). It must be noted that
the corpus column name must be named according to the
following sample :
\itemize{
\item text_item_name : the name of the article/text.
\item text_content : the content of the article/text.
\item text_category : the category/source of the article/text.
}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{D200TechSimilarityTask} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_insight_matrix"></a>}}
\subsection{Method \code{get_insight_matrix()}}{
Create the insight matrix set
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{D200TechSimilarityTask$get_insight_matrix(
  corpus_matrix = self$corpus_matrix,
  set_cat = list(main_cat = "main_cat_name", ref_cat = "ref_cat_name")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{corpus_matrix}}{Like above description, \code{corpus_matrix} is the output
of the private method \code{get_vectorize_matrix}.}

\item{\code{set_cat}}{the \code{list} object contain two following character vector
from \code{text_category} field from raw corpus :
\itemize{
\item main_cat : which types of corpus you want to base on.
\item ref_cat : which types of corpus you want to be recommended.
}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\itemize{
\item a matrix object store in \code{R6_object$insight_matrix_set}.
\item auto create .rds file in insight_matrix_set folder in your
working directory.
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_recommend_result"></a>}}
\subsection{Method \code{get_recommend_result()}}{
Create the recommendation list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{D200TechSimilarityTask$get_recommend_result(
  insight_matrix_set = self$insight_matrix_set,
  topn
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{insight_matrix_set}}{Like above description, corpus_matrix is the output
of the public method \code{get_insight_matrix}. The default value of this
argument is \code{insight_matrix_set} make by latest excutation. You also can use
different version \code{insight_matrix_set} stored in insight_matrix_set folder in
your working directory.}

\item{\code{topn}}{n most relevant item in ref_cat you choice.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The recommendation list will export in recommend_result folder in
your working directory.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_connection"></a>}}
\subsection{Method \code{get_connection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{D200TechSimilarityTask$get_connection(id1, id2)}\if{html}{\out{</div>}}
}

}
}
