% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_project.R
\name{initialize_project}
\alias{initialize_project}
\title{Initialize the working directory.}
\usage{
initialize_project()
}
\description{
\code{initialize_project()} is a side effect function. Must use it at the
beginning of the whole project, it will auto create three folder in your
working directory.
}
\details{
The purpose of each folder:
\itemize{
\item insight_matrix_set : The place store insight_matrix_set's rds file.
\item recommend_result : The place store the result list of recommendation.
}

If each folder has been created in the previous project, the function
will also notify the folder already existed.
}
